rename {Allocator => allocator}.hpp
